pyproject.toml
[project]
name = "atur-projections"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "bokeh>=3.8.2",
    "boto3>=1.42.56",
    "cartopy>=0.25.0",
    "dask>=2026.1.2",
    "distributed>=2026.1.2",
    "fsspec>=2026.2.0",
    "geopandas>=1.1.2",
    "matplotlib>=3.10.8",
    "netcdf4>=1.7.4",
    "numba>=0.64.0",
    "numpy>=2.4.2",
    "s3fs>=0.4.2",
    "scipy>=1.17.1",
    "xarray>=2026.2.0",
    "zarr>=3.1.5",
]
